requirements.txt
ply >= 3.11
pillow >= 9.3.0  # optional

